Remove libigl dependency and implement loadObj

+++ src/mesh.cpp
@@ -2,26 +2,26 @@
 
 #include "mesh.h"
 
-void Mesh::set(const Eigen::MatrixXd &vertices, const Eigen::MatrixXi &faces) {
+void Mesh::set(const Eigen::MatrixXf &vertices, const Eigen::MatrixXi &faces) {
   this->vertices = vertices;
   this->faces = faces;
   this->computeFaceNormals(vertices, faces, this->faceNormals);
   // this->computeVertexNormals(vertices, faces, this->vertexNormals);
 }
 
-void Mesh::computeFaceNormals(const Eigen::MatrixXd &V, const Eigen::MatrixXi &F, Eigen::MatrixXd &N) {
+void Mesh::computeFaceNormals(const Eigen::MatrixXf &V, const Eigen::MatrixXi &F, Eigen::MatrixXf &N) {
 
   std::cout << "Computing face normals .. ";
 
   N.resize(F.rows(), 3);
 
   for(int f=0; f<F.rows(); f++) {
-    const Eigen::Vector3d p0 = V.row(F(f,0));
-    const Eigen::Vector3d p1 = V.row(F(f,1));
-    const Eigen::Vector3d p2 = V.row(F(f,2));
-    const Eigen::Vector3d n0 = (p1 - p0).cross(p2 - p0);
-    const Eigen::Vector3d n1 = (p2 - p1).cross(p0 - p1);
-    const Eigen::Vector3d n2 = (p0 - p2).cross(p1 - p2);
+    const Eigen::Vector3f p0 = V.row(F(f,0));
+    const Eigen::Vector3f p1 = V.row(F(f,1));
+    const Eigen::Vector3f p2 = V.row(F(f,2));
+    const Eigen::Vector3f n0 = (p1 - p0).cross(p2 - p0);
+    const Eigen::Vector3f n1 = (p2 - p1).cross(p0 - p1);
+    const Eigen::Vector3f n2 = (p0 - p2).cross(p1 - p2);
 
     for(int d=0; d<3; d++) {
       N(f,d) = n0(d) + n1(d) + n2(d);
@@ -31,7 +31,7 @@ void Mesh::computeFaceNormals(const Eigen::MatrixXd &V, const Eigen::MatrixXi &F
   std::cout << "done." << std::endl;
 }
 
-void Mesh::computeVertexNormals(const Eigen::MatrixXd &V, const Eigen::MatrixXi &F, Eigen::MatrixXd &N) {
+void Mesh::computeVertexNormals(const Eigen::MatrixXf &V, const Eigen::MatrixXi &F, Eigen::MatrixXf &N) {
 
   std::cout << "Computing vertex normals .. " << std::endl;
 
@@ -39,9 +39,9 @@ void Mesh::computeVertexNormals(const Eigen::MatrixXd &V, const Eigen::MatrixXi 
   N.setZero();
 
   for (int f = 0; f<F.rows(); f++) {
-    Eigen::Vector3d fn = Eigen::Vector3d::Zero();
+    Eigen::Vector3f fn = Eigen::Vector3f::Zero();
     for (int i=0; i<3; ++i) {
-      Eigen::Vector3d v0 = V.col(F(f, i)),
+      Eigen::Vector3f v0 = V.col(F(f, i)),
       v1 = V.col(F(f, (i+1)%3)),
       v2 = V.col(F(f, (i+2)%3)),
       d0 = v1-v0,
